requirements.txt
Babel==1.3
Django==1.8.5
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.7.0
PyYAML==3.11
Pygments==2.0.2
Sphinx==1.2.3
Unidecode==0.04.18
WebOb==1.4.1
WebTest==2.0.17
Werkzeug==0.9.6
Whoosh==2.6.0
Willow==0.2.1
apipkg==1.4
beautifulsoup4==4.4.1
behave==1.2.4
cov-core==1.15.0
coverage==3.7.1
coveralls==0.4.4
django-appconf==1.0.1
django-compressor==1.5
django-crispy-forms==1.5.2
django-debug-toolbar==1.3.0
django-debug-toolbar-template-timings==0.6.4
django-environ==0.4.0
django-extensions==1.3.11
django-extra-views==0.6.4
django-forms-bootstrap==3.0.1
django-haystack==2.3.1
django-jsonfield==0.9.15
django-localflavor==1.0
django-modelcluster==0.6.2
django-oscar==1.1.1
django-oscar-avalara==0.2.1
django-oscar-datacash==0.8.3
django-overextends==0.4.0
django-tables2==1.0.4
django-taggit==0.17.1
django-treebeard==3.0
django-webtest==1.7.7
django-widget-tweaks==1.3
djangorestframework==3.2.4
docopt==0.6.2
docutils==0.12
execnet==1.4.1
factory-boy==2.4.1
flake8==2.4.1
flake8-blind-except==0.1.0
flake8-debugger==1.4.0
html5lib==0.999
ipdb==0.8
ipython==2.3.0
isort==4.2.2
mccabe==0.3.1
mock==1.0.1
nose==1.3.7
parse==1.6.6
parse-type==0.3.4
pep8==1.5.7
phonenumbers==6.3.0
psycopg2==2.6.1
purl==1.1
py==1.4.30
pycountry==1.8
pyflakes==0.8.1
pyprof2calltree==1.3.2
pysolr==3.2.0
pytest==2.7.0
pytest-cache==1.0
pytest-cov==1.8.1
pytest-django==2.8.0
pytest-xdist==1.11
pytz==2015.7
requests==2.7.0
six==1.10.0
sorl-thumbnail==12.2
spec==0.11.1
sphinx-me==0.3
sphinx-rtd-theme==0.1.6
sphinxcontrib-napoleon==0.2.8
sqlparse==0.1.16
stripe==1.27.1
tox==1.8.1
uWSGI==2.0.8
virtualenv==13.1.2
wagtail==1.1
waitress==0.8.10
